refactor: step1 리뷰 반영
- LottoStatistics 의 멤버변수 lottoStatistics의 이름을 lottoHistogram으로 변경
- getDetail 메서드를 toString을 오버라이드하게 변경
- Lotto 클래스의 이름을 LottoService로 변경
- money 변수의 타입을 int에서 long으로 변경
- bonusNumber의 포함여부를 나타내는 boolean 변수의 이름 변경

+++ src/main/java/lotto/domain/LottoDispenser.java
@@ -13,16 +13,16 @@ public class LottoDispenser {
         this.numberSelectStrategy = numberSelectStrategy;
     }
 
-    public LottoTicketsManager getLottoTicket(int money) {
+    public LottoTicketsManager getLottoTicket(long money) {
         List<LottoTicket> lottoTickets = new ArrayList<>();
-        int ticketQuantity = calculateTicketQuantity(money);
+        long ticketQuantity = calculateTicketQuantity(money);
         for (int i = 0; i < ticketQuantity; i++) {
             lottoTickets.add(new LottoTicket(numberSelectStrategy.selectNumbers()));
         }
         return new LottoTicketsManager(lottoTickets);
     }
 
-    private int calculateTicketQuantity(int money) {
+    private long calculateTicketQuantity(long money) {
         return money / LottoTicket.LOTTO_TICKET_PRICE;
     }
 }
